requirements.txt
requests
datetime
python-dotenv
py-clob-client
supabase
backoff
